- Moves note added for the carousel view to the sidebar.
